Change the way we save the VKS data
Instead of using the "output" table, we now save the VKS data in its own
table (VKS_output), putting the answers of each session in its own
column as a giant JSON field. This is done to minimize the row count in
the base, since it's limited by Airtable.

Data is now saved after every word, instead of once at the end of the
session.