[wordcloud] add session query parameter in URL
This gives us a more reliable way of detecting which session the user is
currently doing (before we checked if the session3 information had been
fetched, but this is likely to change and does not work when the
wordcloud page is refreshed.

Also, remove the warning when trying to leave the wordcloud page, since
the posttest is launched using a hard reload and so it triggered the
warning which was confusing for users.