Adapt YesNo to the new Experiment
Also rewrite how data is loaded, still a Work In Progress but YesNo is
already working. Instead of loading all the session data at once we'll
load all the activities info at startup, then load session data when on
the session top page, and load activity data when on the activity's
page.

Also we fetch data on route change, not only on full page load.