Fix activity data loading
loading state was dependent on the order in which functions were called
(starting loading the instructions and then calling the init function of
the module ended up in the (Activity Nothing Nothing) state.

This should be fixed by using dedicated pages for instructions as there
is no reason for the current complexity (adhoc implementation of the
Para library).